Code with new bug, frequencies 28-30Hz
Issue with running motor in this range, not sure cause, switching
methods to try and resolve

+++ CCS_Well_Tester/Nav_Functions.h
@@ -24,8 +24,9 @@ void Straight(	unsigned int Direction,
 void Turn(	unsigned int Direction,
 			unsigned int Left_Profile_ID,
 			unsigned int Right_Profile_ID);
+
 void Create_Nav_Profile(unsigned int Profile_ID,
-						unsigned int Target_Speed,
+						unsigned long Target_Speed,
 						bool Do_Acc,
 						bool Do_Dec);
 void Set_Motor(	unsigned int Motor_ID,
